用户信息修改，头像修改

+++ code/blog-client-ui/src/views/post/new.vue
@@ -19,7 +19,13 @@
               <v-row>
                 <v-col cols="1" align-self="center">标题</v-col>
                 <v-col cols="11">
-                  <v-text-field v-model="form.title" density="compact" variant="outlined" placeholder="一个拥有标识度的标题能够更加的吸引人" hide-details></v-text-field>
+                  <v-text-field
+                    v-model="form.title"
+                    density="compact"
+                    variant="outlined"
+                    placeholder="一个拥有标识度的标题能够更加的吸引人"
+                    hide-details
+                  ></v-text-field>
                 </v-col>
               </v-row>
               <v-row>
@@ -43,19 +49,17 @@
       </template>
       <template #side>
         <v-sheet class="pa-5">
-            <v-card flat>
-                <v-card-title>
-                    发帖提示
-                </v-card-title>
-                <v-card-text>
-                    <ul>
-                        <li>给你的贴子选择正确的分类</li>
-                        <li>鲜明的标题，好的标题更能够吸引人</li>
-                        <li>认真的书写你的正文，真诚是表达的前提</li>
-                        <li>允许转载，请在末尾给出原文链接</li>
-                    </ul>
-                </v-card-text>
-            </v-card>
+          <v-card flat>
+            <v-card-title> 发帖提示 </v-card-title>
+            <v-card-text>
+              <ul>
+                <li>给你的贴子选择正确的分类</li>
+                <li>鲜明的标题，好的标题更能够吸引人</li>
+                <li>认真的书写你的正文，真诚是表达的前提</li>
+                <li>允许转载，请在末尾给出原文链接</li>
+              </ul>
+            </v-card-text>
+          </v-card>
         </v-sheet>
       </template>
     </side-main>
@@ -64,6 +68,7 @@
 
 <script setup>
 import { ref, reactive, onMounted } from 'vue'
+import { useRouter } from 'vue-router'
 
 import Editor from '@toast-ui/editor'
 import '@toast-ui/editor/dist/toastui-editor.css'
@@ -71,29 +76,40 @@ import '@toast-ui/editor/dist/toastui-editor.css'
 import { addArticle, getCaptcha } from '@/http/article'
 import { formdata } from '@/http/request'
 
-
+const router = useRouter()
 let editor
 const captchas = ref()
 let form = reactive({
   title: '',
-  captchas_id: '',
+  captcha_id: '',
   body: '',
 })
 const handlerCategorie = (val) => {
   form.captcha_id = val._id
 }
 const save = () => {
-    const markdownText = editor.getMarkdown()
-    form.body = markdownText
-    const post = JSON.stringify(form)
-    localStorage.setItem('post',post)
-}
-const submit = async () => {
   const markdownText = editor.getMarkdown()
   form.body = markdownText
-  addArticle(form).then(() => {
-    if(localStorage.getItem('post')){
-        localStorage.clear('post')
+  const post = JSON.stringify(form)
+  localStorage.setItem('post', post)
+}
+const submit = async () => {
+  if (form.title) {
+    return
+  }
+  if (form.captcha_id) {
+    return
+  }
+  const markdownText = editor.getMarkdown()
+  form.body = markdownText
+  addArticle(form).then((res) => {
+    if (res.status == 200) {
+      router.push({
+        path: `/posts/${res.data.data._id}`,
+      })
+    }
+    if (localStorage.getItem('post')) {
+      localStorage.clear('post')
     }
   })
 }
@@ -116,9 +132,9 @@ const initEditor = () => {
 }
 
 onMounted(() => {
-    if(localStorage.getItem('post')){
-        form = JSON.parse(localStorage.getItem('post'))
-    }
+  if (localStorage.getItem('post')) {
+    form = JSON.parse(localStorage.getItem('post'))
+  }
   initEditor()
   getCaptcha().then((res) => {
     captchas.value = res.data
@@ -129,8 +145,8 @@ onMounted(() => {
 .containner {
   background: #edf2f7;
   min-height: calc(100vh - 48px);
-  ul{
-    color: rgba(0,0,0,0.5);
+  ul {
+    color: rgba(0, 0, 0, 0.5);
   }
 }
 </style>
